requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

Pillow==9.1.0
pytorch-lightning==1.6.3
tensorboard==2.8.0
torch==1.11.0+cu113
torchmetrics==0.8.2
torchvision==0.12.0+cu113
protobuf==3.20.1
